Fixed reason for deviancy between cython version and python version. Apparently cython does not like the notation 10**(-7), but changing this to 1e-7 fixed the problem. Also changed the type of Q_v_is_zero_two_timesteps in python version from np.int to np.intc.

+++ transition_functions.py
@@ -239,7 +239,7 @@ def I_3(g, Nj, Q_cj, rho_j, rho_a, Ny, Nx, Q_a, Q_th, Q_o, f, a, Q_v):
     '''
     nb_index = [[-1, 0], [-1, 1], [0, 1], [1, 0], [1, -1], [0, -1]]
     nQ_v = np.zeros((Ny, Nx), dtype=np.double, order='C')
-    Q_v_is_zero_two_timesteps = np.zeros((Ny, Nx), dtype=np.int, order='C')
+    Q_v_is_zero_two_timesteps = np.zeros((Ny, Nx), dtype=np.intc, order='C')
     for ii in range(1, Ny-1):
         for jj in range(1, Nx - 1):
             if (Q_th[ii,jj] > 0):
